backend/notification_service/tests/requirements.txt
pytest==7.2.1
pytest-asyncio==0.20.3
aiohttp==3.8.3
pydantic==1.10.4
protobuf==4.21.12
grpcio-tools==1.51.1
